Update Rosetta helper paths
